Sorting, DTOs, User Settings

+++ src/main/java/dev/todoplus/restapi/services/TaskService.java
@@ -1,5 +1,6 @@
 package dev.todoplus.restapi.services;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import dev.todoplus.restapi.data.*;
 import dev.todoplus.restapi.exceptions.*;
 import dev.todoplus.restapi.repositories.SubTaskRepository;
@@ -9,12 +10,16 @@ import dev.todoplus.restapi.repositories.UserRepository;
 import dev.todoplus.restapi.requests.update.UpdateSubTaskRequest;
 import dev.todoplus.restapi.requests.update.UpdateTaskListRequest;
 import dev.todoplus.restapi.requests.update.UpdateTaskRequest;
+import dev.todoplus.restapi.responses.dto.TaskDTO;
 import lombok.val;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * Project: todoplus-restapi
@@ -24,6 +29,8 @@ import java.util.Set;
  */
 @Service
 public class TaskService {
+
+    private final ModelMapper modelMapper = new ModelMapper();
 
     private final TaskListRepository taskListRepository;
     private final TaskRepository taskRepository;
@@ -47,31 +54,35 @@ public class TaskService {
         return taskList;
     }
 
-    public Set<Task> getAllTasks(int taskListId) {
+    public Set<TaskDTO> getAllTasks(int taskListId) {
         val taskList = taskListRepository.getById(taskListId);
-        return taskList.getTasks();
+        return taskList.getTasks().stream().map(task -> modelMapper.map(task, TaskDTO.class)).collect(Collectors.toSet());
     }
 
-    public Set<Task> getMyDayTasks(String username) {
+    public Set<TaskDTO> getMyDayTasks(String username) {
         val user = userRepository.findByUsername(username);
         if (!user.isPresent()) {
             return new HashSet<>();
         }
-        return taskRepository.findAllMyDayTasksByUserId(user.get().getId());
+        return taskRepository.findAllMyDayTasksByUserId(user.get().getId()).stream().map(task -> modelMapper.map(task, TaskDTO.class)).collect(Collectors.toSet());
     }
 
-    public Set<Task> getImportantTasks(String username) {
+    public Set<TaskDTO> getImportantTasks(String username) {
         val user = userRepository.findByUsername(username);
         if (!user.isPresent()) {
             return new HashSet<>();
         }
-        return taskRepository.findAllTasksByUserIdAndImportance(user.get().getId(), Importance.HIGH);
+        return taskRepository.findAllTasksByUserIdAndImportance(user.get().getId(), Importance.HIGH).stream().map(task -> modelMapper.map(task, TaskDTO.class)).collect(Collectors.toSet());
     }
 
     public TaskList createNewTaskList(String username, String displayName) throws UserDoesNotExistException {
         val user = userRepository.findByUsername(username);
         if (!user.isPresent()) {
             throw new UserDoesNotExistException();
+        }
+
+        if (displayName == null || displayName.equals("")) {
+            throw new InvalidNameException();
         }
 
         TaskList taskList = new TaskList();
@@ -83,24 +94,32 @@ public class TaskService {
         return taskListRepository.save(taskList);
     }
 
-    public Task createNewTask(int taskListId, String title) throws TaskListDoesNotExistException {
+    public TaskDTO createNewTask(int taskListId, String title) throws TaskListDoesNotExistException {
         val taskList = taskListRepository.findById(taskListId).orElseThrow(TaskListDoesNotExistException::new);
+
+        if (title == null || title.equals("")) {
+            throw new InvalidNameException();
+        }
 
         Task task = new Task();
         task.setTitle(title);
         task.setStatus(TaskStatus.INPROGRESS);
         task.setDescription(null);
         task.setTaskList(taskList);
-        task.setSort(0);
-        task.setImportance(Importance.NORMAL);
+        task.setSort(taskRepository.getMaxSort(taskListId).orElse(-1) + 1);
+        task.setImportance(Importance.LOW);
         task.setCreateTime(new Date());
 
-        return taskRepository.save(task);
+        return taskToDTO(taskRepository.save(task));
     }
 
-    public Task createNewSubTask(int taskListId, int taskId, String title) throws TaskListDoesNotExistException {
+    public TaskDTO createNewSubTask(int taskListId, int taskId, String title) throws TaskListDoesNotExistException {
         val taskList = taskListRepository.findById(taskListId).orElseThrow(TaskListDoesNotExistException::new);
         val task = taskRepository.findById(taskId).orElseThrow(TaskListDoesNotExistException::new);
+
+        if (title == null || title.equals("")) {
+            throw new InvalidNameException();
+        }
 
         SubTask subTask = new SubTask();
         subTask.setTask(task);
@@ -109,51 +128,75 @@ public class TaskService {
         subTask.setStatus(TaskStatus.INPROGRESS);
 
         subTaskRepository.save(subTask);
-
-        return taskRepository.findById(taskId).orElseThrow(TaskListDoesNotExistException::new);
+        val ret = taskRepository.findById(taskId).orElseThrow(TaskListDoesNotExistException::new);
+        return taskToDTO(ret);
     }
 
-    public Task closeTask(int taskListId, int taskId) throws TaskDoesNotExistException {
+    public TaskDTO closeTask(int taskListId, int taskId) throws TaskDoesNotExistException {
         val task = taskRepository.findById(taskId).orElseThrow(TaskDoesNotExistException::new);
         if (task.getTaskList().getId() == taskListId) {
             task.setCompleteTime(new Date());
             task.setStatus(TaskStatus.COMPLETED);
-
             taskRepository.save(task);
-            return task;
+            taskRepository.moveAllBetween(task.getTaskList().getId(), task.getSort() + 1, 1000000, -1);
+            return taskToDTO(task);
         } else {
             throw new TaskDoesNotExistException();
         }
     }
 
-    public Task closeSubTask(int taskListId, int taskId, int subTaskId) throws SubTaskDoesNotExistException {
-        val subTask = subTaskRepository.findById(taskId).orElseThrow(SubTaskDoesNotExistException::new);
+    public TaskDTO reopenTask(int taskListId, int taskId) throws TaskDoesNotExistException {
+        val task = taskRepository.findById(taskId).orElseThrow(TaskDoesNotExistException::new);
+        if (task.getTaskList().getId() == taskListId) {
+            task.setCompleteTime(null);
+            task.setStatus(TaskStatus.INPROGRESS);
+            task.setSort(taskRepository.getMaxSort(taskListId).orElse(-1) + 1);
+
+            taskRepository.save(task);
+            return taskToDTO(task);
+        } else {
+            throw new TaskDoesNotExistException();
+        }
+    }
+
+    public TaskDTO closeSubTask(int taskListId, int taskId, int subTaskId) throws SubTaskDoesNotExistException {
+        val subTask = subTaskRepository.findById(subTaskId).orElseThrow(SubTaskDoesNotExistException::new);
         if (subTask.getTask().getId() != taskId || subTask.getTask().getTaskList().getId() != taskListId) {
             throw new SubTaskDoesNotExistException();
         }
         subTask.setStatus(TaskStatus.COMPLETED);
         subTaskRepository.save(subTask);
-        return subTask.getTask();
+        return taskToDTO(subTask.getTask());
     }
 
-    public Task addToMyDayList(String username, int taskId) throws TaskDoesNotExistException, NoPermissionsException {
+    public TaskDTO reopenSubTask(int taskListId, int taskId, int subTaskId) throws SubTaskDoesNotExistException {
+        val subTask = subTaskRepository.findById(subTaskId).orElseThrow(SubTaskDoesNotExistException::new);
+        if (subTask.getTask().getId() != taskId || subTask.getTask().getTaskList().getId() != taskListId) {
+            throw new SubTaskDoesNotExistException();
+        }
+        subTask.setStatus(TaskStatus.INPROGRESS);
+        subTaskRepository.save(subTask);
+        return taskToDTO(subTask.getTask());
+    }
+
+    public TaskDTO addToMyDayList(String username, int taskId) throws TaskDoesNotExistException, NoPermissionsException {
         val task = taskRepository.findById(taskId).orElseThrow(TaskDoesNotExistException::new);
         if (!task.getTaskList().getOwner().getUsername().equals(username)) {
             throw new NoPermissionsException();
         }
         task.setMyDay(true);
         taskRepository.save(task);
-        return task;
+        return taskToDTO(task);
     }
 
-    public Task removeFromMyDayList(String username, int taskId) throws TaskDoesNotExistException, NoPermissionsException {
+    public TaskDTO removeFromMyDayList(String username, int taskId) throws TaskDoesNotExistException, NoPermissionsException {
         val task = taskRepository.findById(taskId).orElseThrow(TaskDoesNotExistException::new);
         if (!task.getTaskList().getOwner().getUsername().equals(username)) {
             throw new NoPermissionsException();
         }
         task.setMyDay(false);
         taskRepository.save(task);
-        return task;
+        return taskToDTO(task);
     }
 
     public SubTask createNewSubTask(String username, int taskId, SubTask template) throws TaskListDoesNotExistException {
@@ -193,7 +236,8 @@ public class TaskService {
         return taskList.get();
     }
 
-    public Task updateTask(int taskListId, int taskId, UpdateTaskRequest request) throws TaskDoesNotExistException {
+    @Transactional
+    public TaskDTO updateTask(int taskListId, int taskId, UpdateTaskRequest request) throws TaskDoesNotExistException {
         val task = taskRepository.findById(taskId).orElseThrow(TaskDoesNotExistException::new);
         if (task.getTaskList().getId() != taskListId) {
             throw new TaskDoesNotExistException();
@@ -212,7 +256,18 @@ public class TaskService {
         }
 
         if (request.getSort() != null) {
-            task.setSort(request.getSort());
+            int sourceSort = task.getSort();
+            if(!task.getSort().equals(request.getSort())){
+                if (sourceSort < request.getSort()) { // UP
+                    taskRepository.moveAllBetween(task.getTaskList().getId(), sourceSort + 1, request.getSort(), -1);
+                } else { // DOWN
+                    taskRepository.moveAllBetween(task.getTaskList().getId(), request.getSort(), sourceSort - 1, 1);
+                }
+
+                task.setSort(request.getSort());
+                taskRepository.setSort(taskId, request.getSort());
+                return taskToDTO(task);
+            }
         }
 
         if (request.getDueTime() != null) {
@@ -220,10 +275,21 @@ public class TaskService {
         }
 
         taskRepository.save(task);
-        return task;
+        return taskToDTO(task);
     }
 
-    public Task updateSubTask(int taskListId, int taskId, int subTaskId, UpdateSubTaskRequest request) throws SubTaskDoesNotExistException {
+    public TaskDTO removeDueDate(int taskListId, int taskId) throws TaskDoesNotExistException {
+        val task = taskRepository.findById(taskId).orElseThrow(TaskDoesNotExistException::new);
+        if (task.getTaskList().getId() != taskListId) {
+            throw new TaskDoesNotExistException();
+        }
+        task.setDueTime(null);
+
+        taskRepository.save(task);
+        return taskToDTO(task);
+    }
+
+    public TaskDTO updateSubTask(int taskListId, int taskId, int subTaskId, UpdateSubTaskRequest request) throws SubTaskDoesNotExistException {
         val subTask = subTaskRepository.findById(subTaskId).orElseThrow(SubTaskDoesNotExistException::new);
         if (subTask.getTask().getId() != taskId || subTask.getTask().getTaskList().getId() != taskListId) {
             throw new SubTaskDoesNotExistException();
@@ -238,18 +304,39 @@ public class TaskService {
         }
 
         subTaskRepository.save(subTask);
-        return subTask.getTask();
+        return taskToDTO(subTask.getTask());
     }
 
     public void deleteTaskList(int taskListId) {
         taskListRepository.deleteById(taskListId);
     }
 
-    public void deleteTask(int taskId) {
+
+    public void deleteTask(int taskListId, int taskId) throws TaskDoesNotExistException {
+        val task = taskRepository.findById(taskId).orElseThrow(TaskDoesNotExistException::new);
+        if (task.getTaskList().getId() != taskListId) {
+            throw new TaskDoesNotExistException();
+        }
         taskRepository.deleteById(taskId);
+        taskRepository.moveAllBetween(task.getTaskList().getId(), task.getSort() + 1, 1000000, -1);
     }
 
-    public void deleteSubTask(int subTaskId) {
-        subTaskRepository.deleteById(subTaskId);
+    public TaskDTO deleteSubTaskAndGet(int taskListId, int taskId, int subTaskId) throws SubTaskDoesNotExistException {
+        val subTask = subTaskRepository.findById(subTaskId).orElseThrow(SubTaskDoesNotExistException::new);
+        if (subTask.getTask().getId() != taskId || subTask.getTask().getTaskList().getId() != taskListId) {
+            throw new SubTaskDoesNotExistException();
+        }
+        subTaskRepository.delete(subTask);
+        val task = taskRepository.getById(taskId);
+        return taskToDTO(task);
+    }
+
+    public TaskDTO taskToDTO(Task task) {
+        TaskDTO dto = modelMapper.map(task, TaskDTO.class);
+        dto.setTaskListId(task.getTaskList().getId());
+        if (dto.getSubTasks() == null) {
+            dto.setSubTasks(new HashSet<>());
+        }
+        return dto;
     }
 }
